youtube-comment-analysis/requirements.txt
streamlit==1.32.0
google-api-python-client==2.118.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
nltk==3.8.1
textblob==0.17.1
plotly==5.19.0
wordcloud==1.9.3
python-dotenv==1.0.1 